refactor: Move theming methods to separate partial
