+++ src/Uno.UI/UI/Xaml/DependencyObjectStore.cs
@@ -1387,414 +1387,6 @@ namespace Microsoft.UI.Xaml
 		}
 
 		/// <summary>
-		/// Stores a <see cref="Data.ThemeResourceReference"/> on this object for the given property.
-		/// The reference pins the providing dictionary for O(1) re-resolution on theme change.
-		/// </summary>
-		/// <remarks>
-		/// MUX Reference: CDependencyObject::SetThemeResourceBinding in Theming.cpp
-		/// </remarks>
-		internal void SetThemeResourceBinding(DependencyProperty property, Data.ThemeResourceReference themeRef)
-		{
-			_themeResources ??= new ThemeResourceMap();
-
-			// Resolve the effective precedence using the same logic as SetResourceBinding
-			var precedence = themeRef.Precedence;
-			if (precedence == DependencyPropertyValuePrecedences.Local && _overriddenPrecedences?.Count > 0)
-			{
-				precedence = _overriddenPrecedences.Peek() ?? precedence;
-			}
-
-			_themeResources.Set(property, precedence, themeRef);
-		}
-
-		/// <summary>
-		/// Gets the theme resource references for a specific property, if any.
-		/// Used by animation code to re-resolve ThemeResource values with the
-		/// target element's theme context.
-		/// </summary>
-		internal IEnumerable<Data.ThemeResourceReference>? GetThemeResourceReferences(DependencyProperty property)
-		{
-			return _themeResources?.GetForProperty(property);
-		}
-
-		/// <summary>
-		/// Updates all <see cref="Data.ThemeResourceReference"/> entries using WinUI's
-		/// clear + ancestor-walk + pinned-dict-fallback pattern.
-		/// </summary>
-		/// <remarks>
-		/// MUX Reference: CDependencyObject::UpdateAllThemeReferences in Theming.cpp (lines 260-286)
-		///
-		/// WinUI's algorithm:
-		/// 1. Snapshot all property indices with theme refs
-		/// 2. For each: clear the ref from the map, then re-apply via SetThemeResourceBinding
-		///    which calls UpdateThemeReference(CThemeResource*):
-		///    - If element is active: walk ancestors (FindNextResolvedValueNoRef)
-		///    - Fall back to RefreshValue (pinned dict lookup)
-		/// 3. Get resolved value, SetValue, store ref back in map
-		///
-		/// The clear-first pattern handles cascading changes (e.g., setting Style triggers new theme refs).
-		/// </remarks>
-		internal void UpdateAllThemeReferences(DependencyObject? owner, ThemeWalkResourceCache? cache = null)
-		{
-			if (_themeResources is not { HasEntries: true })
-			{
-				return;
-			}
-
-			// MUX: Theming.cpp:271-276 — snapshot property indices to avoid mutation issues
-			// "Get the properties that have theme refs" into a separate collection before iterating,
-			// because updating one property could cascade and modify _themeResources.
-			var entries = _themeResources.GetAll();
-			var snapshotCount = entries.Count;
-			var snapshot = new ThemeResourceMap.Entry[snapshotCount];
-			for (var i = 0; i < snapshotCount; i++)
-			{
-				snapshot[i] = entries[i];
-			}
-
-			for (var i = 0; i < snapshot.Length; i++)
-			{
-				var property = snapshot[i].Property;
-				var precedence = snapshot[i].Precedence;
-				var themeRef = snapshot[i].Reference;
-				try
-				{
-					// MUX: Theming.cpp:297-298 — clear the theme ref first.
-					// "Just continue to the next one if the ref isn't in the map now,
-					//  which means it no longer applies to this element."
-					if (_themeResources.Get(property, precedence) != themeRef)
-					{
-						continue; // ref was removed by a cascading update
-					}
-					_themeResources.Clear(property, precedence);
-
-					// MUX: Theming.cpp:315-346 — UpdateThemeReference(CThemeResource*)
-					object? newValue = null;
-					bool resolved = false;
-
-					// Phase A: Ancestor walk (WinUI: FindNextResolvedValueNoRef)
-					// If element is active, walk ancestor ResourceDictionaries to find
-					// the resource. This handles re-parenting correctly.
-					if (owner is FrameworkElement { IsLoaded: true })
-					{
-						var dicts = GetResourceDictionaries(includeAppResources: false);
-						foreach (var dict in dicts)
-						{
-							if (dict.TryGetValue(themeRef.ResourceKey, out var ancestorValue, shouldCheckSystem: false))
-							{
-								themeRef.LastResolvedValue = ancestorValue;
-								newValue = ancestorValue;
-								resolved = true;
-								break;
-							}
-						}
-					}
-
-					// Phase B: Pinned dict fallback (WinUI: themeResource->RefreshValue())
-					// MUX: Theming.cpp:340-342 — "Call refresh if we're in a theme walk
-					// or the ref has been updated in the past and the value wasn't updated
-					// already by the tree lookup above."
-					if (!resolved)
-					{
-						newValue = themeRef.RefreshValue(owner, cache);
-					}
-
-					var convertedValue = BindingPropertyHelper.Convert(property.Type, newValue);
-
-					// Skip deferred entries that haven't been resolved yet (null target dict,
-					// null last value). They will be resolved by the _resourceBindings fallback
-					// path during loading.
-					if (convertedValue is null && themeRef.LastResolvedValue is null)
-					{
-						// MUX: Theming.cpp:397 — store ref back even if we skip the value set
-						_themeResources.Set(property, precedence, themeRef);
-						continue;
-					}
-
-					if (themeRef.SetterBindingPath is { } bindingPath)
-					{
-						try
-						{
-							_isSettingPersistentResourceBinding = true;
-							bindingPath.Value = convertedValue;
-						}
-						finally
-						{
-							_isSettingPersistentResourceBinding = false;
-						}
-					}
-					else
-					{
-						SetValue(property, convertedValue, precedence, isPersistentResourceBinding: true);
-					}
-
-					// MUX: Theming.cpp:397 — store the reference back in the map
-					_themeResources.Set(property, precedence, themeRef);
-				}
-				catch (Exception e)
-				{
-					if (this.Log().IsEnabled(Uno.Foundation.Logging.LogLevel.Warning))
-					{
-						this.Log().Warn($"Failed to update theme resource binding for {property.Name}", e);
-					}
-					// Restore the ref even on failure to avoid losing it
-					_themeResources.Set(property, precedence, themeRef);
-				}
-			}
-		}
-
-		/// <summary>
-		/// Do a tree walk to find the correct values of StaticResource and ThemeResource assignations.
-		/// </summary>
-		internal void UpdateResourceBindings(ResourceUpdateReason updateReason, FrameworkElement? resourceContextProvider = null, ResourceDictionary? containingDictionary = null)
-		{
-			if (updateReason == ResourceUpdateReason.None)
-			{
-				throw new ArgumentException();
-			}
-
-			// Phase 1: Update theme resources via ancestor-walk + pinned-dictionary path
-			// MUX Reference: CDependencyObject::UpdateAllThemeReferences() in Theming.cpp
-			if ((updateReason & ResourceUpdateReason.ThemeResource) != 0)
-			{
-				UpdateAllThemeReferences(ActualInstance, ThemeWalkResourceCache.Instance);
-			}
-
-			ResourceDictionary[]? dictionariesInScope = null;
-
-			if ((updateReason & ResourceUpdateReason.ThemeResource) != 0 &&
-				_properties.HasBindings)
-			{
-				dictionariesInScope = GetResourceDictionaries(includeAppResources: false, resourceContextProvider, containingDictionary).ToArray();
-				for (var i = dictionariesInScope.Length - 1; i >= 0; i--)
-				{
-					ResourceResolver.PushSourceToScope(dictionariesInScope[i]);
-				}
-
-				_properties.UpdateBindingExpressions();
-
-				for (int i = 0; i < dictionariesInScope.Length; i++)
-				{
-					ResourceResolver.PopSourceFromScope();
-				}
-			}
-
-			// Phase 2: Update remaining resource bindings (StaticResourceLoading, HotReload) via tree walk
-			if (_resourceBindings?.HasBindings == true)
-			{
-				dictionariesInScope ??= GetResourceDictionaries(includeAppResources: false, resourceContextProvider, containingDictionary).ToArray();
-
-				var bindings = _resourceBindings.GetAllBindings();
-				foreach (var binding in bindings)
-				{
-					InnerUpdateResourceBindings(updateReason, dictionariesInScope, binding.Property, binding.Binding);
-				}
-			}
-
-			UpdateChildResourceBindings(updateReason, resourceContextProvider);
-		}
-
-		/// <remarks>
-		/// This method contains or is called by a try/catch containing method and
-		/// can be significantly slower than other methods as a result on WebAssembly.
-		/// See https://github.com/dotnet/runtime/issues/56309
-		/// </remarks>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		private void InnerUpdateResourceBindings(ResourceUpdateReason updateReason, ResourceDictionary[] dictionariesInScope, DependencyProperty property, ResourceBinding binding)
-		{
-			try
-			{
-				InnerUpdateResourceBindingsUnsafe(updateReason, dictionariesInScope, property, binding);
-			}
-			catch (Exception e)
-			{
-				if (this.Log().IsEnabled(Uno.Foundation.Logging.LogLevel.Warning))
-				{
-					this.Log().Warn($"Failed to update binding, target may have been disposed", e);
-				}
-			}
-		}
-
-		/// <remarks>
-		/// This method contains or is called by a try/catch containing method and
-		/// can be significantly slower than other methods as a result on WebAssembly.
-		/// See https://github.com/dotnet/runtime/issues/56309
-		/// </remarks>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		private void InnerUpdateResourceBindingsUnsafe(ResourceUpdateReason updateReason, ResourceDictionary[] dictionariesInScope, DependencyProperty property, ResourceBinding binding)
-		{
-			if ((binding.UpdateReason & updateReason) == ResourceUpdateReason.None)
-			{
-				// If the reason for the update doesn't match the reason(s) that the binding was created for, don't update it
-				return;
-			}
-
-			if ((updateReason & ResourceUpdateReason.ResolvedOnLoading) != 0)
-			{
-				// Add the current dictionaries to the resolver scope,
-				// this allows for StaticResource.ResourceKey to resolve properly
-
-				for (var i = dictionariesInScope.Length - 1; i >= 0; i--)
-				{
-					ResourceResolver.PushSourceToScope(dictionariesInScope[i]);
-				}
-			}
-
-			var wasSet = false;
-			foreach (var dict in dictionariesInScope)
-			{
-				if (dict.TryGetValue(binding.ResourceKey, out var value, out var providingDict, shouldCheckSystem: false))
-				{
-					wasSet = true;
-					SetResourceBindingValue(property, binding, value);
-
-					// Pin the providing dictionary in the _themeResources entry (if any)
-					// so that subsequent theme changes can re-resolve from it directly.
-					if (providingDict is not null && _themeResources is not null)
-					{
-						var themeRef = _themeResources.Get(property, binding.Precedence);
-						themeRef?.SetTargetDictionary(providingDict);
-						if (themeRef is not null)
-						{
-							themeRef.LastResolvedValue = value;
-						}
-					}
-
-					break;
-				}
-			}
-
-			if (!wasSet)
-			{
-				if (ResourceResolver.TryTopLevelRetrieval(binding.ResourceKey, binding.ParseContext, out var value))
-				{
-					SetResourceBindingValue(property, binding, value);
-				}
-			}
-
-			if ((updateReason & ResourceUpdateReason.ResolvedOnLoading) != 0)
-			{
-				foreach (var dict in dictionariesInScope)
-				{
-					ResourceResolver.PopSourceFromScope();
-				}
-			}
-		}
-
-		[UnconditionalSuppressMessage("Trimming", "IL2067", Justification = "normal flow of operation")]
-		[UnconditionalSuppressMessage("Trimming", "IL2072", Justification = "normal flow of operation")]
-		private void SetResourceBindingValue(DependencyProperty property, ResourceBinding binding, object? value)
-		{
-			var convertedValue = BindingPropertyHelper.Convert(property.Type, value);
-			if (binding.SetterBindingPath != null)
-			{
-				try
-				{
-					_isSettingPersistentResourceBinding = binding.IsPersistent;
-					binding.SetterBindingPath.Value = convertedValue;
-				}
-				finally
-				{
-					_isSettingPersistentResourceBinding = false;
-				}
-			}
-			else
-			{
-				SetValue(property, convertedValue, binding.Precedence, isPersistentResourceBinding: binding.IsPersistent);
-			}
-		}
-
-		private bool _isUpdatingChildResourceBindings;
-
-		private void UpdateChildResourceBindings(ResourceUpdateReason updateReason, FrameworkElement? resourceContextProvider = null)
-		{
-			if (_isUpdatingChildResourceBindings)
-			{
-				// Some DPs might be creating reference cycles, so we make sure not to enter an infinite loop.
-				return;
-			}
-
-			if ((updateReason & ResourceUpdateReason.PropagatesThroughTree) != ResourceUpdateReason.None)
-			{
-				try
-				{
-					InnerUpdateChildResourceBindings(updateReason, resourceContextProvider);
-				}
-				finally
-				{
-					_isUpdatingChildResourceBindings = false;
-				}
-
-				if (ActualInstance is IThemeChangeAware themeChangeAware)
-				{
-					// Call OnThemeChanged after bindings of descendants have been updated
-					themeChangeAware.OnThemeChanged();
-				}
-
-				_properties.OnThemeChanged();
-			}
-		}
-
-		/// <remarks>
-		/// This method contains or is called by a try/catch containing method and
-		/// can be significantly slower than other methods as a result on WebAssembly.
-		/// See https://github.com/dotnet/runtime/issues/56309
-		/// </remarks>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		private void InnerUpdateChildResourceBindings(ResourceUpdateReason updateReason, FrameworkElement? resourceContextProvider = null)
-		{
-			_isUpdatingChildResourceBindings = true;
-
-			foreach (var propertyDetail in _properties.GetAllDetails())
-			{
-				if (propertyDetail == null
-					|| propertyDetail == _properties.DataContextPropertyDetails)
-				{
-					continue;
-				}
-
-				var propertyValue = GetValue(propertyDetail);
-
-				if (propertyValue is IEnumerable<DependencyObject> dependencyObjectCollection &&
-					// Try to avoid enumerating collections that shouldn't be enumerated, since we may be encountering user-defined values. This may need to be refined to somehow only consider values coming from the framework itself.
-					(propertyValue is ICollection || propertyValue is DependencyObjectCollectionBase)
-				)
-				{
-					foreach (var innerValue in dependencyObjectCollection)
-					{
-						UpdateResourceBindingsIfNeeded(innerValue, updateReason, resourceContextProvider);
-					}
-				}
-
-				if (propertyValue is IAdditionalChildrenProvider updateable)
-				{
-					foreach (var innerValue in updateable.GetAdditionalChildObjects())
-					{
-						UpdateResourceBindingsIfNeeded(innerValue, updateReason, resourceContextProvider);
-					}
-				}
-
-				if (propertyValue is DependencyObject dependencyObject)
-				{
-					UpdateResourceBindingsIfNeeded(dependencyObject, updateReason, resourceContextProvider);
-				}
-			}
-		}
-
-		private void UpdateResourceBindingsIfNeeded(DependencyObject dependencyObject, ResourceUpdateReason updateReason, FrameworkElement? resourceContextProvider = null)
-		{
-			// propagate to non-FE DO
-			if (dependencyObject is not IFrameworkElement && dependencyObject is IDependencyObjectStoreProvider storeProvider)
-			{
-				storeProvider.Store.UpdateResourceBindings(
-					updateReason,
-					// when propagating to non-FE, we need to inject a FE as the resource context
-					resourceContextProvider: resourceContextProvider ?? ActualInstance as FrameworkElement
-				);
-			}
-		}
-
-		/// <summary>
 		/// Returns all ResourceDictionaries in scope using the visual tree, from nearest to furthest.
 		/// </summary>
 		internal IEnumerable<ResourceDictionary> GetResourceDictionaries(
